fix: resolved conflicts done
